+++ lib/chips_choice.dart
@@ -1,7 +1,7 @@
 library chips_choice;
 
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart' show FilterChip, ThemeData, Theme;
 
 /// Easy way to provide a single or multiple choice chips.
 class ChipsChoice<T> extends StatelessWidget {
@@ -15,6 +15,9 @@ class ChipsChoice<T> extends StatelessWidget {
   /// Builder for custom choice item
   final ChipsChoiceBuilder<T> itemBuilder;
 
+  /// Whether the chips is wrapped or scrollable
+  final bool isWrapped;
+
   /// List padding
   final EdgeInsetsGeometry padding;
 
@@ -24,6 +27,7 @@ class ChipsChoice<T> extends StatelessWidget {
   final ChipsChoiceChanged<List<T>> _onChangedMultiple;
   final bool _isMultiChoice;
 
+  /// Costructor for single choice
   ChipsChoice.single({
     Key key,
     @required T value,
@@ -31,8 +35,10 @@ class ChipsChoice<T> extends StatelessWidget {
     @required ChipsChoiceChanged<T> onChanged,
     this.itemConfig = const ChipsChoiceItemConfig(),
     this.itemBuilder,
+    this.isWrapped = false,
     this.padding,
   }) : assert(onChanged != null),
+    assert(isWrapped != null),
     _isMultiChoice = false,
     _value = value,
     _values = null,
@@ -40,6 +46,7 @@ class ChipsChoice<T> extends StatelessWidget {
     _onChangedSingle = onChanged,
     super(key: key);
 
+  /// Constructor for multiple choice
   ChipsChoice.multiple({
     Key key,
     @required List<T> value,
@@ -47,18 +54,20 @@ class ChipsChoice<T> extends StatelessWidget {
     @required ChipsChoiceChanged<List<T>> onChanged,
     this.itemConfig = const ChipsChoiceItemConfig(),
     this.itemBuilder,
+    this.isWrapped = false,
     this.padding,
-  }) : assert(onChanged != null && value != null),
+  }) : assert(onChanged != null),
+    assert(isWrapped != null),
     _isMultiChoice = true,
     _value = null,
-    _values = value,
+    _values = value ?? [],
     _onChangedSingle = null,
     _onChangedMultiple = onChanged,
     super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return itemConfig.isWrapped == true
+    return isWrapped == true
       ? _listWrapped()
       : _listScrollable();
   }
@@ -80,31 +89,31 @@ class ChipsChoice<T> extends StatelessWidget {
         horizontal: 15.0,
       ),
       child: Wrap(
-        spacing: itemConfig.wrappedSpacing, // gap between adjacent chips
-        runSpacing: itemConfig.wrappedRunSpacing, // gap between lines
+        spacing: itemConfig.spacing, // gap between adjacent chips
+        runSpacing: itemConfig.runSpacing, // gap between lines
         children: _choiceItems,
       ),
     );
   }
 
   List<Widget> get _choiceItems {
-    return List<Widget>.generate(
-      options.length,
-      (i) => _choiceItemBuilder(i),
-    ).where((item) => item != null).toList();
+    return List<Widget>
+      .generate(options.length, _choiceItemsGenerator)
+      .where((e) => e != null).toList();
   }
 
-  Widget _choiceItemBuilder(int i) {
+  Widget _choiceItemsGenerator(int i) {
     ChipsChoiceOption<T> item = options[i];
     bool selected = _isMultiChoice
       ? _values.contains(item.value)
       : _value == item.value;
     return item.hidden == false
-      ? itemBuilder?.call(item, itemConfig, selected) ?? _ChipsChoiceItem(
+      ? itemBuilder?.call(item, selected, _onSelect) ?? _ChipsChoiceItem(
           item: item,
           config: itemConfig,
           selected: selected,
           onSelect: _onSelect,
+          isWrapped: isWrapped,
         )
       : null;
   }
@@ -124,12 +133,14 @@ class ChipsChoice<T> extends StatelessWidget {
   }
 }
 
+// Default choice item
 class _ChipsChoiceItem<T> extends StatelessWidget {
 
   final ChipsChoiceOption<T> item;
   final ChipsChoiceItemConfig config;
   final Function onSelect;
   final bool selected;
+  final bool isWrapped;
 
   _ChipsChoiceItem({
     Key key,
@@ -137,6 +148,7 @@ class _ChipsChoiceItem<T> extends StatelessWidget {
     @required this.config,
     @required this.onSelect,
     @required this.selected,
+    this.isWrapped = false,
   }) : super(key: key);
 
   @override
@@ -146,7 +158,7 @@ class _ChipsChoiceItem<T> extends StatelessWidget {
     Color _unselectedColor = config.unselectedColor ?? _theme.unselectedWidgetColor;
 
     return Padding(
-      padding: config.margin,
+      padding: config.margin ?? isWrapped ? EdgeInsets.all(0) : EdgeInsets.all(5),
       child: FilterChip(
         label: Text(
           item.label,
@@ -166,8 +178,8 @@ class _ChipsChoiceItem<T> extends StatelessWidget {
         pressElevation: config.pressElevation,
         shadowColor: _unselectedColor,
         selectedShadowColor: _selectedColor,
-        backgroundColor: Colors.transparent,
-        selectedColor: Colors.transparent,
+        backgroundColor: Color(0x00000000),
+        selectedColor: Color(0x00000000),
         checkmarkColor: _selectedColor,
         showCheckmark: config.showCheckmark == true,
         selected: selected,
@@ -179,13 +191,25 @@ class _ChipsChoiceItem<T> extends StatelessWidget {
   }
 }
 
+/// Choice option
 class ChipsChoiceOption<T> {
+
+  /// Value to return
   final T value;
+
+  /// Represent as primary text
   final String label;
+
+  /// Represent as chips avatar
   final Widget avatar;
+
+  /// Whether the choice is disabled or not
   final bool disabled;
+
+  /// Whether the choice is hidden or displayed
   final bool hidden;
 
+  /// Default Constructor
   ChipsChoiceOption({
     @required this.value,
     @required this.label,
@@ -195,13 +219,14 @@ class ChipsChoiceOption<T> {
   }) : assert(disabled != null),
     assert(hidden != null);
 
-  static List<ChipsChoiceOption<R>> listFrom<S, R>({
-    @required List<S> source,
-    @required ChipsChoiceOptionProp<S, R> value,
-    @required ChipsChoiceOptionProp<S, String> label,
-    ChipsChoiceOptionProp<S, Widget> avatar,
-    ChipsChoiceOptionProp<S, bool> disabled,
-    ChipsChoiceOptionProp<S, bool> hidden,
+  /// Helper to create option list from any list
+  static List<ChipsChoiceOption<R>> listFrom<R, E>({
+    @required List<E> source,
+    @required _ChipsChoiceOptionProp<E, R> value,
+    @required _ChipsChoiceOptionProp<E, String> label,
+    _ChipsChoiceOptionProp<E, Widget> avatar,
+    _ChipsChoiceOptionProp<E, bool> disabled,
+    _ChipsChoiceOptionProp<E, bool> hidden,
   }) => source
     .asMap()
     .map((index, item) => MapEntry(index, ChipsChoiceOption<R>(
@@ -216,37 +241,64 @@ class ChipsChoiceOption<T> {
     .cast<ChipsChoiceOption<R>>();
 }
 
+/// Choice item configuration
 class ChipsChoiceItemConfig {
-  final bool isWrapped;
+
+  /// whether the chips use checkmark or not
   final bool showCheckmark;
+
+  /// Selected item color
   final Color selectedColor;
+
+  /// Unselected item color
   final Color unselectedColor;
+
+  /// choice item margin
   final EdgeInsetsGeometry margin;
-  final double wrappedSpacing;
-  final double wrappedRunSpacing;
+
+  /// How much space to place between children in a run in the main axis.
+  /// When [ChipsChoice.isWrapped] is [true]
+  final double spacing;
+
+  /// How much space to place between the runs themselves in the cross axis.
+  /// When [ChipsChoice.isWrapped] is [true]
+  final double runSpacing;
+
+  /// Chips elevation
   final double elevation;
+
+  /// Longpress chips elevation
   final double pressElevation;
+
+  /// Chips shape builder
   final ChipsChoiceShapeBuilder shapeBuilder;
 
+  /// Default Constructor
   const ChipsChoiceItemConfig({
-    this.isWrapped = false,
     this.showCheckmark = true,
     this.selectedColor,
     this.unselectedColor,
-    this.margin = const EdgeInsets.all(5),
-    this.wrappedSpacing = 10.0,
-    this.wrappedRunSpacing = 0,
+    this.margin,
+    this.spacing = 10.0,
+    this.runSpacing = 0,
     this.elevation = 0,
     this.pressElevation = 0,
     this.shapeBuilder,
   });
 }
 
+/// Builder for option prop
+typedef R _ChipsChoiceOptionProp<T, R>(int index, T item);
+
+/// Builder for chips shape border
 typedef ShapeBorder ChipsChoiceShapeBuilder(bool selected);
-typedef R ChipsChoiceOptionProp<T, R>(int index, T item);
+
+/// Callback when the value changed
 typedef void ChipsChoiceChanged<T>(T values);
+
+/// Builder for custom choice item
 typedef Widget ChipsChoiceBuilder<T>(
   ChipsChoiceOption<T> item,
-  ChipsChoiceItemConfig config,
-  bool selected
+  bool selected,
+  Function(T value, bool selected) onSelect,
 );